backend/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# Core AI/ML
google-genai
google-generativeai

# Web framework
aiohttp
beautifulsoup4
fastapi
feedparser
gpt-researcher
numpy
pydantic
pydantic-settings
python-dateutil
python-dotenv
requests
scikit-learn
tavily-python
uvicorn[standard]

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Background tasks
apscheduler>=3.10.0
